Create file set-timeline.ts for reuse
