+++ practice/src/scripts/controllers/task.controler.ts
@@ -17,7 +17,6 @@ class TaskController {
         this.taskView.setTaskFilter(this.handleFilterTask);
         this.taskView.setCheckAllToggleTask(this.handleCheckAllToggleTask);
         this.taskView.setClearCompletedHandler(this.handleClearCompleted);
-        this.handleReloadWindows();
     }
 
     public init = (): void =>  {
@@ -31,20 +30,6 @@ class TaskController {
     handleCheckAllToggleTask = (): void => {
         this.taskModel.checkAllToggleTask(this.taskView.renderTasks);
         this.taskView.updateClearCompletedButtonVisibility();
-    };
-
-    handleReloadWindows = (): void => {
-        const status = window.location.hash;
-        let elements: NodeList = querySelectorAll(".task-filter-item > a");
-
-        elements.forEach(function (element) {
-            if (element instanceof HTMLElement) {
-                element.classList.remove("active-btn");
-                if (element.dataset.action === status.slice(1, status.length)) {
-                        element.classList.add("active-btn");
-                }
-            }
-        });
     };
 
     handleFilterTask = (actionFilter: string): void => {
