+++ practice/src/scripts/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './display-element';
 export * from './dom-element';
 export * from './messages';
+export * from './set-timeline';
